pyproject.toml
[project]
name = "symprosys-task"
version = "0.1.0"
description = "The project is about a small ecommerce platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.135.3",
    "pydantic==2.12.5",
    "psycopg2-binary==2.9.11",
    "structlog==25.5.0",
    "uvicorn==0.42.0",
    "SQLAlchemy==2.0.48",
    "alembic==1.18.4",
    "python-dotenv==1.2.2",
    "PyJWT==2.12.1"
]
